requirements.txt
click
rackspacesdk